Header and carousel customization additions; code clean-up
Enable custom header images through WordPress theme admin, instead of
hard-coding the Blue & White header.  The new default header image is
completely transparent, both to emphasize the spot color and because I'm
not very good at graphic design.

Allow the user to specify a heading for the featured article carousel on
the home page, instead of always using "In This Issue".  The surrounding
<section> element has had its id attribute changed accordingly, from
"home-inthisissue" to "home-featured".

Eliminate various hard-coded mentions of Blue & White and dead code,
including a version of jQuery that's already shipped with WordPress.